requirements.txt
django-cors-headers
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv